# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/messages": {
      "get": {
        "operationId": "MessageController_getMessages",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": false,
            "in": "query",
            "description": "Filter by chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Filter by sender. A phone also matches messages from a lid that resolves to it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history"
          }
        },
        "summary": "Get message history for a session",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/messages/{chatId}/history": {
      "get": {
        "description": "Reads messages directly from the WhatsApp client for the given chat, bypassing the local DB. Useful for retrieving messages that arrived before the gateway was started.",
        "operationId": "MessageController_getChatHistory",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID (e.g. 1234567890@c.us or groupId@g.us)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeMedia",
            "required": false,
            "in": "query",
            "description": "When true, downloads media (base64) for messages that have it. Slower; default false.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "deep",
            "required": false,
            "in": "query",
            "description": "When true, raises the limit ceiling from 100 to 2000 for reaching further back in history (whatsapp-web.js only; loads earlier messages on demand). Forces metadata-only (includeMedia is ignored). Large/slow requests may increase WhatsApp rate-limiting risk; default false.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat history (most recent messages)"
          }
        },
        "summary": "Fetch chat history live from WhatsApp",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/{chatId}/{messageId}/reactions": {
      "get": {
        "operationId": "MessageController_getReactions",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID containing the message",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "description": "Message ID to get reactions for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reactions with senders"
          }
        },
        "summary": "Get reactions for a specific message",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/messages/{waMessageId}": {
      "get": {
        "operationId": "MessageLookupController_getMessage",
        "parameters": [
          {
            "name": "waMessageId",
            "required": true,
            "in": "path",
            "description": "WhatsApp message id, e.g. false_120363400287378579@g.us_AC60FA8596471EA47F3EBA73160CEE76_2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Message not found"
          }
        },
        "summary": "Get a single message by its WhatsApp message id (no session required)",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.8.17",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}